+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Add a tutorial for exporting logs to S3.
+- Add team planeteers Slack alert routing (`#alert-planeteers`, `#alert-planeteers-test` on non-stable installations).
 
 ## [0.77.0] - 2026-09-07